requirements.txt
# Smart AI Chatbot - Requirements
# Install with: pip install -r requirements.txt

# Web Framework
streamlit>=1.28.0

# xAI Grok API (OpenAI-compatible SDK)
openai>=1.0.0

# LangChain for RAG
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.2.0

# Vector Store
faiss-cpu>=1.7.4

# Embeddings
sentence-transformers>=2.2.0

# PDF Processing
pypdf>=3.17.0

# Web Search
duckduckgo-search>=4.0.0

# Environment Variables
python-dotenv>=1.0.0

# Additional utilities
numpy>=1.24.0
